requirements.txt
allure-pytest==2.6.0
allure-python-commons==2.8.18
docker==4.3.1
pytest==6.0.2
pytest-sugar==0.9.4
requests==2.24.0
loguru==0.5.3
win32-setctime==1.0.3